Project 2/requirements.txt
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-ollama>=0.1.0

fastapi[standard]>=0.113.0, <0.114.0
pydantic>=2.7.0, <3.0.0
scikit-learn>=1.6.1
httpx>=0.28.1
